+++ CHANGELOG.md
@@ -15,6 +15,11 @@
 - **buession-security-captcha：** AliYunCaptchaClient 构造函数参数 regionId 为 null 或空字符串时，不再抛出异常；而使用默认值
 
 
+### 🐞 Bug 修复
+
+- **buession-security-core：** 修复 SameSite LAX 值
+
+
 ---
 
 
@@ -27,7 +32,7 @@
 
 ### 🐞 Bug 修复
 
-- **buession-captcha：** 修复极验 v3 版本生成签名错误的问题
+- **buession-security-captcha：** 修复极验 v3 版本生成签名错误的问题
 
 
 ---